finalized map implementation, moved scripts and fixed precompile issues

+++ db/migrate/20171025152137_create_posts.rb
@@ -6,10 +6,11 @@ class CreatePosts < ActiveRecord::Migration[5.1]
       t.string :title
       t.string :image
       t.string :location
-      t.integer :latitude
-      t.integer :longitude
+      t.float :latitude
+      t.float :longitude
 
       t.timestamps
     end
   end
 end
+
